backend/products/pyproject.toml
[tool.poetry]
name = "products"
version = "0.1.0"
description = "Microservice for products"
authors = ["Bogdan Loginov <bogdanloginov31@gmail.com>"]
license = "MIT Licence"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.3"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
django-admin-interface = "^0.28.6"
celery = "^5.3.6"
djangorestframework = "^3.15.1"
gunicorn = "^21.2.0"
requests = "^2.31.0"
daphne = "^4.1.2"
channels = "^4.1.0"

[tool.poetry.dev-dependencies]
flake8 = "^7.0.0"
poetry-plugin-export = "^1.7.1"
rapidfuzz = "^3.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
